renaming current iam endpoints, queries, and types
