+++ packages/api-v4/src/iam/iam.ts
@@ -1,19 +1,19 @@
 import { BETA_API_ROOT } from '../constants';
 import Request, { setData, setMethod, setURL } from '../request';
 
-import type { IamAccountPermissions, IamUserPermissions } from './types';
+import type { IamAccountRoles, IamUserRoles } from './types';
 
 /**
- * getUserPermissions
+ * getUserRoles
  *
  * Returns the full permissions structure for this User. This includes all entities on
  * the Account alongside what level of access this User has to each of them.
  *
- * @param username { number } the username to look up.
+ * @param username { string } the username to look up.
  *
  */
-export const getUserPermissions = (username: string) =>
-  Request<IamUserPermissions>(
+export const getUserRoles = (username: string) =>
+  Request<IamUserRoles>(
     setURL(
       `${BETA_API_ROOT}/iam/users/${encodeURIComponent(
         username,
@@ -21,20 +21,18 @@ export const getUserPermissions = (username: string) =>
     ),
     setMethod('GET'),
   );
+
 /**
- * updateUserPermissions
+ * updateUserRoles
  *
- * Update the permissions a User has.
+ * Update the roles a User has.
  *
- * @param username { number } ID of the client to be viewed.
- * @param data { object } the Permissions object to update.
+ * @param username { string } username of the user to be updated.
+ * @param data { object } the Roles object to update.
  *
  */
-export const updateUserPermissions = (
-  username: string,
-  data: IamUserPermissions,
-) =>
-  Request<IamUserPermissions>(
+export const updateUserRoles = (username: string, data: IamUserRoles) =>
+  Request<IamUserRoles>(
     setURL(
       `${BETA_API_ROOT}/iam/users/${encodeURIComponent(
         username,
@@ -45,13 +43,13 @@ export const updateUserPermissions = (
   );
 
 /**
- * getAccountPermissions
+ * getAccountRoles
  *
- * Return all permissions for account.
+ * Return all roles for account.
  *
  */
-export const getAccountPermissions = () => {
-  return Request<IamAccountPermissions>(
+export const getAccountRoles = () => {
+  return Request<IamAccountRoles>(
     setURL(`${BETA_API_ROOT}/iam/role-permissions`),
     setMethod('GET'),
   );
