+++ packages/manager/src/features/IAM/Roles/Roles.tsx
@@ -3,18 +3,18 @@ import React from 'react';
 
 import { RolesTable } from 'src/features/IAM/Roles/RolesTable/RolesTable';
 import { mapAccountPermissionsToRoles } from 'src/features/IAM/Shared/utilities';
-import { useAccountPermissions } from 'src/queries/iam/iam';
+import { useAccountRoles } from 'src/queries/iam/iam';
 
 export const RolesLanding = () => {
-  const { data: accountPermissions, isLoading } = useAccountPermissions();
+  const { data: accountRoles, isLoading } = useAccountRoles();
 
   const { roles } = React.useMemo(() => {
-    if (!accountPermissions) {
+    if (!accountRoles) {
       return { roles: [] };
     }
-    const roles = mapAccountPermissionsToRoles(accountPermissions);
+    const roles = mapAccountPermissionsToRoles(accountRoles);
     return { roles };
-  }, [accountPermissions]);
+  }, [accountRoles]);
 
   if (isLoading) {
     return <CircleProgress />;
