+++ packages/manager/src/features/IAM/Shared/AssignedRolesTable/ChangeRoleDrawer.tsx
@@ -12,13 +12,13 @@ import { useParams } from 'react-router-dom';
 
 import { Link } from 'src/components/Link';
 import {
-  useAccountPermissions,
-  useAccountUserPermissions,
-  useAccountUserPermissionsMutation,
+  useAccountRoles,
+  useUserRoles,
+  useUserRolesMutation,
 } from 'src/queries/iam/iam';
 
 import { AssignedPermissionsPanel } from '../AssignedPermissionsPanel/AssignedPermissionsPanel';
-import { getAllRoles, getRoleByName, updateUserRoles } from '../utilities';
+import { changeUserRole, getAllRoles, getRoleByName } from '../utilities';
 
 import type { DrawerModes, EntitiesOption, ExtendedRoleView } from '../types';
 import type { RolesType } from '../utilities';
@@ -34,13 +34,12 @@ export const ChangeRoleDrawer = ({ mode, onClose, open, role }: Props) => {
   const theme = useTheme();
   const { username } = useParams<{ username: string }>();
 
-  const { data: accountPermissions, isLoading: accountPermissionsLoading } =
-    useAccountPermissions();
+  const { data: accountRoles, isLoading: accountPermissionsLoading } =
+    useAccountRoles();
 
-  const { data: assignedRoles } = useAccountUserPermissions(username ?? '');
+  const { data: assignedRoles } = useUserRoles(username ?? '');
 
-  const { mutateAsync: updateUserPermissions } =
-    useAccountUserPermissionsMutation(username);
+  const { mutateAsync: updateUserRoles } = useUserRolesMutation(username);
 
   const formattedAssignedEntities: EntitiesOption[] = React.useMemo(() => {
     if (!role || !role.entity_names || !role.entity_ids) {
@@ -55,14 +54,14 @@ export const ChangeRoleDrawer = ({ mode, onClose, open, role }: Props) => {
 
   // filtered roles by entity_type and access
   const allRoles = React.useMemo(() => {
-    if (!accountPermissions) {
+    if (!accountRoles) {
       return [];
     }
 
-    return getAllRoles(accountPermissions).filter(
+    return getAllRoles(accountRoles).filter(
       (el) => el.entity_type === role?.entity_type && el.access === role?.access
     );
-  }, [accountPermissions, role]);
+  }, [accountRoles, role]);
 
   const {
     control,
@@ -83,12 +82,12 @@ export const ChangeRoleDrawer = ({ mode, onClose, open, role }: Props) => {
 
   // Get the selected role based on the `selectedOptions`
   const selectedRole = React.useMemo(() => {
-    if (!selectedOptions || !accountPermissions) {
+    if (!selectedOptions || !accountRoles) {
       return null;
     }
 
-    return getRoleByName(accountPermissions, selectedOptions.value);
-  }, [selectedOptions, accountPermissions]);
+    return getRoleByName(accountRoles, selectedOptions.value);
+  }, [selectedOptions, accountRoles]);
 
   const onSubmit = async (data: { roleName: RolesType }) => {
     if (role?.name === data.roleName.label) {
@@ -100,14 +99,14 @@ export const ChangeRoleDrawer = ({ mode, onClose, open, role }: Props) => {
       const newRole = data.roleName.label;
       const access = data.roleName.access;
 
-      const updatedUserRoles = updateUserRoles({
+      const updatedUserRoles = changeUserRole({
         access,
         assignedRoles,
         initialRole,
         newRole,
       });
 
-      await updateUserPermissions(updatedUserRoles);
+      await updateUserRoles(updatedUserRoles);
 
       handleClose();
     } catch (errors) {
