+++ packages/manager/src/features/IAM/Shared/AssignedRolesTable/UnassignRoleConfirmationDialog.tsx
@@ -4,10 +4,7 @@ import React from 'react';
 import { useParams } from 'react-router-dom';
 
 import { ConfirmationDialog } from 'src/components/ConfirmationDialog/ConfirmationDialog';
-import {
-  useAccountUserPermissions,
-  useAccountUserPermissionsMutation,
-} from 'src/queries/iam/iam';
+import { useUserRoles, useUserRolesMutation } from 'src/queries/iam/iam';
 
 import { deleteUserRole } from '../utilities';
 
@@ -29,11 +26,11 @@ export const UnassignRoleConfirmationDialog = (props: Props) => {
   const {
     error,
     isPending,
-    mutateAsync: updateUserPermissions,
+    mutateAsync: updateUserRoles,
     reset,
-  } = useAccountUserPermissionsMutation(username);
+  } = useUserRolesMutation(username);
 
-  const { data: assignedRoles } = useAccountUserPermissions(username ?? '');
+  const { data: assignedRoles } = useUserRoles(username ?? '');
 
   const onClose = () => {
     reset(); // resets the error state of the useMutation
@@ -50,7 +47,7 @@ export const UnassignRoleConfirmationDialog = (props: Props) => {
       initialRole,
     });
 
-    await updateUserPermissions(updatedUserRoles);
+    await updateUserRoles(updatedUserRoles);
 
     enqueueSnackbar(`Role ${role?.name} has been deleted successfully.`, {
       variant: 'success',
