+++ packages/manager/src/features/IAM/Shared/RemoveAssignmentConfirmationDialog/RemoveAssignmentConfirmationDialog.tsx
@@ -4,10 +4,7 @@ import React from 'react';
 import { useParams } from 'react-router-dom';
 
 import { ConfirmationDialog } from 'src/components/ConfirmationDialog/ConfirmationDialog';
-import {
-  useAccountUserPermissions,
-  useAccountUserPermissionsMutation,
-} from 'src/queries/iam/iam';
+import { useUserRoles, useUserRolesMutation } from 'src/queries/iam/iam';
 
 import { deleteUserEntity } from '../utilities';
 
@@ -29,11 +26,11 @@ export const RemoveAssignmentConfirmationDialog = (props: Props) => {
   const {
     error,
     isPending,
-    mutateAsync: updateUserPermissions,
+    mutateAsync: updateUserRoles,
     reset,
-  } = useAccountUserPermissionsMutation(username);
+  } = useUserRolesMutation(username);
 
-  const { data: assignedRoles } = useAccountUserPermissions(username ?? '');
+  const { data: assignedRoles } = useUserRoles(username ?? '');
 
   const onClose = () => {
     reset(); // resets the error state of the useMutation
@@ -52,7 +49,7 @@ export const RemoveAssignmentConfirmationDialog = (props: Props) => {
       entity_type
     );
 
-    await updateUserPermissions({
+    await updateUserRoles({
       ...assignedRoles,
       entity_access: updatedUserEntityRoles,
     });
