+++ packages/manager/src/features/IAM/Shared/utilities.test.ts
@@ -1,8 +1,9 @@
-import { accountPermissionsFactory } from 'src/factories/accountPermissions';
-import { userPermissionsFactory } from 'src/factories/userPermissions';
+import { accountRolesFactory } from 'src/factories/accountRoles';
+import { userRolesFactory } from 'src/factories/userRoles';
 
 import {
   changeRoleForEntity,
+  changeUserRole,
   deleteUserEntity,
   deleteUserRole,
   getAllRoles,
@@ -13,7 +14,6 @@ import {
   mergeAssignedRolesIntoExistingRoles,
   partition,
   toEntityAccess,
-  updateUserRoles,
 } from './utilities';
 
 import type { EntitiesRole, ExtendedRoleView, RoleView } from './types';
@@ -23,7 +23,7 @@ import type { EntityAccess } from '@linode/api-v4';
 const accountAccess = 'account_access';
 const entityAccess = 'entity_access';
 
-const accountPermissions = accountPermissionsFactory.build({
+const accountPermissions = accountRolesFactory.build({
   account_access: [
     {
       roles: [
@@ -62,7 +62,7 @@ const accountPermissions = accountPermissionsFactory.build({
   ],
 });
 
-const userPermissions = userPermissionsFactory.build({
+const userPermissions = userRolesFactory.build({
   account_access: ['account_linode_admin', 'linode_creator'],
   entity_access: [
     {
@@ -196,7 +196,7 @@ describe('getRoleByName', () => {
   });
 });
 
-describe('updateUserRoles', () => {
+describe('changeUserRole', () => {
   it('should return an object of updated users roles with resource access', () => {
     const expectedRoles = {
       account_access: ['account_linode_admin', 'linode_creator'],
@@ -212,7 +212,7 @@ describe('updateUserRoles', () => {
     const initialRole = 'linode_contributor';
     const newRole = 'linode_admin';
     expect(
-      updateUserRoles({
+      changeUserRole({
         access: entityAccess,
         assignedRoles: userPermissions,
         initialRole,
@@ -243,7 +243,7 @@ describe('deleteUserRole', () => {
   it('should return an object of updated users roles with resource access', () => {
     const initialRole = 'linode_contributor';
 
-    const userPermissions = userPermissionsFactory.build();
+    const userPermissions = userRolesFactory.build();
 
     const expectedRoles = {
       account_access: [
@@ -344,7 +344,7 @@ describe('changeRoleForEntity', () => {
   });
 
   it('should return an object of updated users roles with entity access when changing role from "linode_contributor" to "linode_viewer"', () => {
-    const userPermissions = userPermissionsFactory.build({
+    const userPermissions = userRolesFactory.build({
       account_access: ['account_linode_admin', 'linode_creator'],
       entity_access: [
         {
@@ -388,7 +388,7 @@ describe('changeRoleForEntity', () => {
   });
 
   it('should return an object of updated users roles with entity access', () => {
-    const userPermissions = userPermissionsFactory.build({
+    const userPermissions = userRolesFactory.build({
       account_access: ['account_linode_admin', 'linode_creator'],
       entity_access: [
         {
