+++ packages/manager/src/features/IAM/Users/UserEntities/AssignedEntitiesTable.test.tsx
@@ -3,7 +3,7 @@ import userEvent from '@testing-library/user-event';
 import React from 'react';
 
 import { accountEntityFactory } from 'src/factories/accountEntities';
-import { userPermissionsFactory } from 'src/factories/userPermissions';
+import { userRolesFactory } from 'src/factories/userRoles';
 import { makeResourcePage } from 'src/mocks/serverHandlers';
 import { renderWithTheme } from 'src/utilities/testHelpers';
 
@@ -11,14 +11,14 @@ import { AssignedEntitiesTable } from '../../Users/UserEntities/AssignedEntities
 
 const queryMocks = vi.hoisted(() => ({
   useAccountEntities: vi.fn().mockReturnValue({}),
-  useAccountUserPermissions: vi.fn().mockReturnValue({}),
+  useUserRoles: vi.fn().mockReturnValue({}),
 }));
 
 vi.mock('src/queries/iam/iam', async () => {
   const actual = await vi.importActual<any>('src/queries/iam/iam');
   return {
     ...actual,
-    useAccountUserPermissions: queryMocks.useAccountUserPermissions,
+    useUserRoles: queryMocks.useUserRoles,
   };
 });
 
@@ -40,7 +40,7 @@ const mockEntities = [
 
 describe('AssignedEntitiesTable', () => {
   it('should display no roles text if there are no roles assigned to user', async () => {
-    queryMocks.useAccountUserPermissions.mockReturnValue({
+    queryMocks.useUserRoles.mockReturnValue({
       data: {},
     });
 
@@ -50,8 +50,8 @@ describe('AssignedEntitiesTable', () => {
   });
 
   it('should display roles and menu when data is available', async () => {
-    queryMocks.useAccountUserPermissions.mockReturnValue({
-      data: userPermissionsFactory.build(),
+    queryMocks.useUserRoles.mockReturnValue({
+      data: userRolesFactory.build(),
     });
 
     queryMocks.useAccountEntities.mockReturnValue({
@@ -73,8 +73,8 @@ describe('AssignedEntitiesTable', () => {
   });
 
   it('should display empty state when no roles match filters', async () => {
-    queryMocks.useAccountUserPermissions.mockReturnValue({
-      data: userPermissionsFactory.build(),
+    queryMocks.useUserRoles.mockReturnValue({
+      data: userRolesFactory.build(),
     });
 
     queryMocks.useAccountEntities.mockReturnValue({
@@ -92,8 +92,8 @@ describe('AssignedEntitiesTable', () => {
   });
 
   it('should filter roles based on search query', async () => {
-    queryMocks.useAccountUserPermissions.mockReturnValue({
-      data: userPermissionsFactory.build(),
+    queryMocks.useUserRoles.mockReturnValue({
+      data: userRolesFactory.build(),
     });
 
     queryMocks.useAccountEntities.mockReturnValue({
@@ -111,8 +111,8 @@ describe('AssignedEntitiesTable', () => {
   });
 
   it('should filter roles based on selected resource type', async () => {
-    queryMocks.useAccountUserPermissions.mockReturnValue({
-      data: userPermissionsFactory.build(),
+    queryMocks.useUserRoles.mockReturnValue({
+      data: userRolesFactory.build(),
     });
 
     queryMocks.useAccountEntities.mockReturnValue({
