+++ packages/manager/src/features/IAM/Users/UserEntities/AssignedEntitiesTable.tsx
@@ -18,7 +18,7 @@ import { TableRowLoading } from 'src/components/TableRowLoading/TableRowLoading'
 import { TableSortCell } from 'src/components/TableSortCell';
 import { usePagination } from 'src/hooks/usePagination';
 import { useAccountEntities } from 'src/queries/entities/entities';
-import { useAccountUserPermissions } from 'src/queries/iam/iam';
+import { useUserRoles } from 'src/queries/iam/iam';
 
 import { ENTITIES_TABLE_PREFERENCE_KEY } from '../../Shared/constants';
 import { RemoveAssignmentConfirmationDialog } from '../../Shared/RemoveAssignmentConfirmationDialog/RemoveAssignmentConfirmationDialog';
@@ -92,7 +92,7 @@ export const AssignedEntitiesTable = () => {
     data: assignedRoles,
     error: assignedRolesError,
     isLoading: assignedRolesLoading,
-  } = useAccountUserPermissions(username ?? '');
+  } = useUserRoles(username ?? '');
 
   const { filterableOptions, roles } = React.useMemo(() => {
     if (!assignedRoles || !entities) {
