+++ packages/manager/src/features/IAM/Users/UserEntities/ChangeRoleForEntityDrawer.tsx
@@ -12,9 +12,9 @@ import { useParams } from 'react-router-dom';
 
 import { Link } from 'src/components/Link';
 import {
-  useAccountPermissions,
-  useAccountUserPermissions,
-  useAccountUserPermissionsMutation,
+  useAccountRoles,
+  useUserRoles,
+  useUserRolesMutation,
 } from 'src/queries/iam/iam';
 
 import { AssignedPermissionsPanel } from '../../Shared/AssignedPermissionsPanel/AssignedPermissionsPanel';
@@ -43,24 +43,23 @@ export const ChangeRoleForEntityDrawer = ({
   const theme = useTheme();
   const { username } = useParams<{ username: string }>();
 
-  const { data: accountPermissions, isLoading: accountPermissionsLoading } =
-    useAccountPermissions();
+  const { data: accountRoles, isLoading: accountPermissionsLoading } =
+    useAccountRoles();
 
-  const { data: assignedRoles } = useAccountUserPermissions(username ?? '');
+  const { data: assignedRoles } = useUserRoles(username ?? '');
 
-  const { mutateAsync: updateUserPermissions } =
-    useAccountUserPermissionsMutation(username);
+  const { mutateAsync: updateUserRoles } = useUserRolesMutation(username);
 
   // filtered roles by entity_type and access
   const allRoles = React.useMemo(() => {
-    if (!accountPermissions) {
+    if (!accountRoles) {
       return [];
     }
 
-    return getAllRoles(accountPermissions).filter(
+    return getAllRoles(accountRoles).filter(
       (el) => el.entity_type === role?.entity_type && el.access === role?.access
     );
-  }, [accountPermissions, role]);
+  }, [accountRoles, role]);
 
   const {
     control,
@@ -81,12 +80,12 @@ export const ChangeRoleForEntityDrawer = ({
 
   // Get the selected role based on the `selectedOptions`
   const selectedRole = React.useMemo(() => {
-    if (!selectedOptions || !accountPermissions) {
+    if (!selectedOptions || !accountRoles) {
       return null;
     }
 
-    return getRoleByName(accountPermissions, selectedOptions.value);
-  }, [selectedOptions, accountPermissions]);
+    return getRoleByName(accountRoles, selectedOptions.value);
+  }, [selectedOptions, accountRoles]);
 
   const onSubmit = async (data: { roleName: ExtendedEntityRole }) => {
     if (role?.role_name === data.roleName.label) {
@@ -107,7 +106,7 @@ export const ChangeRoleForEntityDrawer = ({
         newRole
       );
 
-      await updateUserPermissions({
+      await updateUserRoles({
         ...assignedRoles!,
         entity_access: updatedEntityRoles,
       });
