+++ packages/manager/src/features/IAM/Users/UserEntities/UserEntities.tsx
@@ -11,7 +11,7 @@ import React from 'react';
 import { useParams } from 'react-router-dom';
 
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
-import { useAccountUserPermissions } from 'src/queries/iam/iam';
+import { useUserRoles } from 'src/queries/iam/iam';
 
 import {
   ERROR_STATE_TEXT,
@@ -28,7 +28,8 @@ export const UserEntities = () => {
     data: assignedRoles,
     isLoading,
     error: assignedRolesError,
-  } = useAccountUserPermissions(username ?? '');
+  } = useUserRoles(username ?? '');
+
   const { error } = useAccountUser(username ?? '');
 
   const hasAssignedRoles = assignedRoles
