+++ packages/manager/src/features/IAM/hooks/useIsIAMEnabled.ts
@@ -1,5 +1,5 @@
 import { useFlags } from 'src/hooks/useFlags';
-import { useAccountPermissions } from 'src/queries/iam/iam';
+import { useAccountRoles } from 'src/queries/iam/iam';
 
 /**
  * Hook to determine if the IAM feature is enabled for the current user.
@@ -8,10 +8,10 @@ import { useAccountPermissions } from 'src/queries/iam/iam';
  */
 export const useIsIAMEnabled = () => {
   const flags = useFlags();
-  const { data: rolePermissions } = useAccountPermissions(flags.iam?.enabled);
+  const { data: accountRoles } = useAccountRoles(flags.iam?.enabled);
 
-  const hasAccountAccess = rolePermissions?.account_access?.length;
-  const hasEntityAccess = rolePermissions?.entity_access?.length;
+  const hasAccountAccess = accountRoles?.account_access?.length;
+  const hasEntityAccess = accountRoles?.entity_access?.length;
 
   return {
     isIAMBeta: flags.iam?.beta,
