+++ packages/manager/src/queries/iam/iam.ts
@@ -1,25 +1,21 @@
-import { updateUserPermissions } from '@linode/api-v4';
+import { updateUserRoles } from '@linode/api-v4';
 import { queryPresets } from '@linode/queries';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 
 import { iamQueries } from './queries';
 
-import type {
-  APIError,
-  IamAccountPermissions,
-  IamUserPermissions,
-} from '@linode/api-v4';
+import type { APIError, IamAccountRoles, IamUserRoles } from '@linode/api-v4';
 
-export const useAccountUserPermissions = (username?: string) => {
-  return useQuery<IamUserPermissions, APIError[]>({
+export const useUserRoles = (username?: string) => {
+  return useQuery<IamUserRoles, APIError[]>({
     ...iamQueries.user(username ?? '')._ctx.permissions,
     refetchOnMount: 'always',
     enabled: Boolean(username),
   });
 };
 
-export const useAccountPermissions = (enabled = true) => {
-  return useQuery<IamAccountPermissions, APIError[]>({
+export const useAccountRoles = (enabled = true) => {
+  return useQuery<IamAccountRoles, APIError[]>({
     ...iamQueries.permissions,
     ...queryPresets.oneTimeFetch,
     ...queryPresets.noRetry,
@@ -27,12 +23,12 @@ export const useAccountPermissions = (enabled = true) => {
   });
 };
 
-export const useAccountUserPermissionsMutation = (username: string) => {
+export const useUserRolesMutation = (username: string) => {
   const queryClient = useQueryClient();
-  return useMutation<IamUserPermissions, APIError[], IamUserPermissions>({
-    mutationFn: (data) => updateUserPermissions(username, data),
+  return useMutation<IamUserRoles, APIError[], IamUserRoles>({
+    mutationFn: (data) => updateUserRoles(username, data),
     onSuccess: (role) => {
-      queryClient.setQueryData<IamUserPermissions>(
+      queryClient.setQueryData<IamUserRoles>(
         iamQueries.user(username)._ctx.permissions.queryKey,
         role
       );
